+++ CHANGELOG.md
@@ -11,6 +11,8 @@
   - Unignore files. (!1033, !1054)
 - Serialization
   - Make timestamp in filename for `QuantumDevice.to_json_file` optional. (!1063)
+- Qblox backend
+  - Cleanup after hardware config redesign. Add `ChannelPath` class. (!1066)
 
 ## 0.21.1 (2024-08-16)